Migrate to use XP Compiler
