+++ composer.json
@@ -7,6 +7,7 @@
   "keywords": ["xp"],
   "require" : {
     "xp-framework/core": "^9.0",
+    "xp-framework/compiler": "^2.10",
     "xp-framework/rdbms": "^11.0",
     "xp-forge/inject": "^4.2",
     "xp-forge/rest-api": "^0.8",
