+++ src/main/php/de/thekid/shorturl/api/Administration.php
@@ -6,18 +6,18 @@ use web\rest\Response;
 class Administration extends Handler {
 
   /** Returns all URLs */
-  #[@get('/'), @$user: value, @$request: request]
-  public function all(string $user, Request $request): Response {
+  <<get('/')>>
+  public function all(<<value>> string $user, <<request>> Request $request): Response {
     $pagination= $this->paging()->on($request);
     return $pagination->paginate($this->urls->all(
-      $pagination->start(),
+      $pagination->start() ?: 0,
       $pagination->limit() + 1
     ));
   }
 
   /** Deletes a URL by a given ID */
-  #[@delete('/{id}'), @$user: value]
-  public function delete(string $user, string $id): Response {
+  <<delete('/{id}')>>
+  public function delete(<<value>> string $user, string $id): Response {
     $this->urls->remove($id);
     return Response::noContent();
   }
